No longer exclude legendary items from being handled with a strategy, implement legendary strategy
